toefl-writing-main/requirements.txt
Flask==2.3.2
openai==0.27.6
python-dotenv==1.0.0
flask-cors==3.0.10